net: extract message processing into net_processing.{h,cpp}
Third step of the CConnman backport series (issue #2558). Bitcoin keeps P2P
message processing in net_processing.{h,cpp}; Gridcoin still had it inline in
main.cpp. Create the translation unit and move the three message functions --
file-static ProcessMessage, ProcessMessages, and SendMessages -- together with
the helpers used only by them, byte-identical:

- Orphan-tx subsystem: mapOrphanTransactions / mapOrphanTransactionsByPrev and
  AddOrphanTx / EraseOrphanTx / LimitOrphanTxSize.
- The static wallet-scan GetTransaction(uint256, CWalletTx&) and Inventory()
  wrappers, whose only callers were the trio. Their sibling wrappers
  (EraseFromWallets / SetBestChain / PrintWallets / UpdatedTransaction) keep
  callers in main.cpp and stay there; the moved pair reaches the registered
  wallets through the existing setpwalletRegistered externs in main.h.
- AlreadyHave().

The moved code reaches shared state through existing headers (main.h via
checkpoints.h, net.h). Two main.cpp file-globals it touches are not header
declared, so net_processing.cpp redeclares them extern: mapAlerts/cs_mapAlerts
(the idiom rpc/net.cpp uses) and cPeerBlockCounts (still owned by main.cpp,
which reads it from GetNumBlocksOfPeers). The ProcessMessages/SendMessages
declarations move from main.h to net_processing.h; net.cpp (ThreadMessageHandler2,
their only caller) includes the new header.

No behavioral change. dos_tests.cpp reaches the orphan subsystem via its own
extern declarations, which relink against net_processing.o unchanged. The
Gridcoin wire-protocol handlers are preserved verbatim: the ARIES/VERSION
alias, the tx path with contract dispatch, alert, scraperindex, and part.
